I made some changes for you to look at.
The library key was for my users collections but I got stuck there are decided to move on to more pressing matters. So for now I just commented all mention of the library key out. Also, I noticed my API key was disabled it. So I enabled it and the urls are running fine on postman now. But I'm still not retrieving any data and I don't understand the errors I'm getting. Here is the code with all my changes.